requirements.txt
streamlit
streamlit-option-menu
psycopg2-binary
